GAP-6 + GAP-10 + GAP-12: put currency case, cross-merchant uniques and the micros/cents boundary under the schema and the type system
These three move together because they cannot move apart: the currency
CHECK turns a previously-silent lowercase write into a hard error, so the
lowercase constants and comparisons had to be corrected in the same
change or the CHECK would have taken production down.

GAP-6 / or#832 — currency codes were app-validated only, and normalisation
was missing at the WRITE boundary. Measured before a dev DB was torn down:
payments.currency was 100% lowercase. The root cause was pkg/catalog's
manifest loader lower-casing on the way in; nothing normalised on the way
out. Chose a DB CHECK over boundary normalisation: the write boundary is
not one place — 16 currency columns are written from dozens of INSERTs
across modules, workers and importers — and a normaliser would be one
more thing to remember. The CHECK is the only point every write passes.
It constrains CASE and rough shape (upper, 3..12, or a qualified
`slug/name` custom unit), not membership; membership stays in the Go
registry where per-currency scale lives and can change without a
migration. Length is 3..12 rather than exactly 3 because a price may be
denominated in a stablecoin (USDC/USDG).

Migration 0020 applies it by ITERATING information_schema for
column_name='currency', not from a hand-written list — a hand-written
list is exactly how 0005's payment_settlement_events escaped the RLS
guard (GAP-4). A test asserts the loop stays derived and that nothing
later drops a constraint.

The trap the CHECK would otherwise have sprung, fixed in the same commit:
subscriptions.CurrencyUSD/EUR, solana/support.go's `currency != "usd"`
comparisons, ccbill's FlexForm code table (whose inverse lookup returned
lowercase straight into a payments row), and the fx quote currencies are
all canonical UPPER now. Lowercase survives only where a rail WIRE demands
it — Stripe's API and the fawazahmed0 FX endpoint — and those two sites
say so. Catalog credit units normalise the same way (unqualified = a
built-in currency = upper) because they reach a ledger currency column via
DepositCatalogCreditPurchase.

GAP-10 / SEC-24 item 4 — checkout_sessions' (rail, reference) and
(rail, transaction_id) were the only cross-merchant uniques in the schema.
Under RLS the conflicting row is invisible, so one merchant's session
surfaces in another's tenant as an opaque insert failure: an existence
oracle and a claim-squat. Both are merchant- and PSP-scoped now, following
0017's COALESCE-the-nullable-psp technique so one TOTAL index keeps
per-PSP scope with no disjoint-partial hole.

TestUniqueIndexesAreMerchantScoped makes that permanent, and found a
SECOND instance the audit missed: catalog_drift_events is merchant_id NOT
NULL and RLS-protected, yet its open-drift unique spanned merchants — and
an NMI plan id is only unique within a gateway account, so two merchants
genuinely can collide. Also scoped in 0020. The five remaining unscoped
uniques are exempted BY NAME with reasons (three key a merchant-owned FK;
a Solana PDA is globally unique on-chain; a PSP gateway account belongs to
one merchant by design).

GAP-12 — scoped to the highest-value boundary: the unit CONVERTERS.
CentsToMicros / MicrosToCentsCeil / MicrosToCentsExact are the only places
a money value changes unit and they were int64 -> int64, so the type
system said nothing at the exact point where saying the wrong thing costs
10 000x. They are typed now, along with ParseDecimalToCents and the
Format* display helpers, and the type propagated outward on its own into
the NMI plan pusher, the CCBill amount-tolerance comparison, the
chargeback matcher and the admin refund path.

DELIBERATELY LEFT (named in docs/invariants.md): DB columns stay bare
bigint; most struct fields stay int64; money.NativeToRailMinor keeps an
int64 input because its argument is "internal units at the CURRENCY's
registered scale" (JPY is 10^4), which is not always micros — typing it
Micros would be a lie.

+++ internal/integrations/ccbill/subscription_management.go
@@ -347,7 +347,7 @@ func (c *DataLinkClient) refundForm(action, subscriptionID, transactionID string
 		form.Set("transactionId", transactionID)
 	}
 	if amountCents > 0 {
-		form.Set("amount", moneyutil.FormatCentsDecimal(int64(amountCents)))
+		form.Set("amount", moneyutil.FormatCentsDecimal(amountCents))
 	}
 	return form
 }
